# public/swagger/json/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "UNOARITMETICO",
        "description": "API unoaritmetico, to use endpoints that need authentication use the 'Set Globally' function first",
        "contact": {
            "email": "walook.com.mx"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "unoaritmetico.com",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Login in the api",
                "description": "Login",
                "operationId": "loginAccount",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Account to register",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Token was given",
                        "schema": {
                            "$ref": "#/definitions/LoginResponse"
                        }
                    },
                    "405": {
                        "description": "Invalid Method",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "422": {
                        "description": "Invalid Fields",
                        "schema": {
                            "$ref": "#/definitions/Validation"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Register a account",
                "description": "Register a account",
                "operationId": "registerAccount",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Account to register",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/RegisterUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Account was registered"
                    },
                    "405": {
                        "description": "Invalid Method",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "422": {
                        "description": "Invalid Fields",
                        "schema": {
                            "$ref": "#/definitions/Validation"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/auth/activate": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Activate account",
                "description": "Activates a account",
                "operationId": "activateUser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "Token given by the mail",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deliverys",
                        "schema": {
                            "$ref": "#/definitions/ActivateUserResponse"
                        }
                    },
                    "400": {
                        "description": "Request format isn't valid",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "405": {
                        "description": "Invalid Method",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "RegisterUserRequest": {
            "properties": {
                "first_name": {
                    "description": "",
                    "type": "string",
                    "example": "John"
                },
                "last_name": {
                    "description": "",
                    "type": "string",
                    "example": "Papa"
                },
                "email": {
                    "description": "",
                    "type": "string",
                    "format": "email",
                    "example": "unoaritmeticoapi@gmail.com"
                },
                "password": {
                    "description": "",
                    "type": "string",
                    "format": "password",
                    "example": "kittens"
                },
                "password_confirmation": {
                    "description": "",
                    "type": "string",
                    "format": "password",
                    "example": "kittens"
                }
            }
        },
...
        "LoginRequest": {
            "properties": {
                "email": {
                    "description": "",
                    "type": "string",
                    "example": "unoaritmeticoapi@gmail.com"
                },
                "password": {
                    "description": "",
                    "type": "string",
                    "example": "kittens"
                }
            }
        },
...
        "ActivateUserResponse": {
            "properties": {
                "meta": {
                    "description": "",
                    "type": "object",
                    "$ref": "#/definitions/DefaultMeta"
                },
                "data": {
                    "description": "",
                    "type": "array",
                    "$ref": "#/definitions/UserItem"
                }
            }
        },
        "LoginResponse": {
            "properties": {
                "meta": {
                    "description": "",
                    "type": "object",
                    "$ref": "#/definitions/TokenMeta"
                },
                "data": {
                    "description": "",
                    "type": "object",
                    "$ref": "#/definitions/UserItem"
                }
            }
        },
...
        "UserItem": {
            "properties": {
                "id": {
                    "description": "",
                    "type": "integer",
                    "example": 1
                },
                "first_name": {
                    "description": "",
                    "type": "string",
                    "example": "John"
                },
                "last_name": {
                    "description": "",
                    "type": "string",
                    "example": "Papa"
                },
                "email": {
                    "description": "",
                    "type": "string",
                    "format": "email",
                    "example": "unoaritmeticoapi@gmail.com"
                },
                "role": {
                    "description": "",
                    "type": "string",
                    "example": "TEACHER"
                }
            }
        },
        "Error": {
            "properties": {
                "message": {
                    "description": "",
                    "type": "string",
                    "example": "Bad Request"
                },
                "status_code": {
                    "description": "",
                    "type": "integer",
                    "format": "int64",
                    "example": 400
                }
            }
        },
        "Validation": {
            "properties": {
                "error": {
                    "description": "",
                    "type": "array",
                    "$ref": "#/definitions/ValidationItem"
                }
            }
        },
        "ValidationItem": {
            "properties": {
                "field": {
                    "description": "",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "DefaultMeta": {},
...
        "TokenMeta": {
            "properties": {
                "token": {
                    "description": "",
                    "type": "string",
                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIsImp0aSI6IjI3MzgzZTUwZDBiMGZmMWZlNGFkMDg3YzRmNDY4ZTIw"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiAuth": {
            "type": "oauth2",
            "name": "Authorization",
            "flow": "password",
            "tokenUrl": "unoaritmetico.com/api",
            "scopes": {
                "read:users": "Consult userss",
                "write:users": "Modify users"
            }
        }
    }
}